Tidying up classifier file, NB and LSTM files
