+++ backend/object_generation.py
@@ -11,10 +11,14 @@ import cProfile
 
 # libraries made by us for this project
 import nlp.SentimentClassifier as VAD
+import LSTMClassifiers
 from ShapeRepresentation.shape import generate_cylinder
 
 # sentiment classifier with 3 dimensions (Valence, Arousal, Dominance)
 vad_classifier = VAD.VADClassifier('nlp/emobank.csv')
+
+# above is outdated. should be:
+#vad_classifier = LSTMClassifiers.LSTMClassifier()
 
 
 def graphs(points, fig_id):
